Fixing Routes and Renders UI
