+++ DiceGame/views/games.ejs
@@ -99,7 +99,7 @@
                       <td
                         class="px-5 py-5 border-b border-gray-200 bg-white text-sm"
                       >
-                        <a href="/api/<%=element._id%>" class="">
+                        <a href="/game/<%=element._id%>" class="">
                           <span class= " whitespace-no-wrap bg-blue-300 hover:bg-blue-500  border border-blue-400 rounded-full text-gray-600 hover:text-white p-2" >Info</span>
                         </a>
                       </td>
@@ -108,8 +108,7 @@
                   </tbody>
                 </table>
                 <% }else { %>
-                  <div c
-                  lass="bg-red-200 p-3">There no data</div>
+                  <div class="bg-red-200 p-3">There no data</div>
                 <% } %> 
               </div>
             </div>
