+++ DiceGame/views/index.ejs
@@ -31,7 +31,7 @@
             <div class="hidden sm:block sm:ml-6">
               <div class="flex space-x-4">
                 <a
-                  href="api/createGame"
+                  href="/createGame"
                   class="text-white hover:bg-green-800 hover:text-white px-10 py-6 rounded-md text-sm font-medium"
                   aria-current="page"
                   >Create Game</a
@@ -50,7 +50,7 @@
                 >
 
                 <a
-                  href="/api/games"
+                  href="/game/all"
                   class="text-white hover:bg-green-800 hover:text-white px-10 py-6 rounded-md text-sm font-medium"
                   >Games</a
                 >
@@ -59,7 +59,6 @@
           </div>
         </div>
       </div>
-
       <!-- Mobile menu, show/hide based on menu state. -->
       <div class="sm:hidden" id="mobile-menu">
         <div class="px-2 pt-2 pb-3 space-y-1">
@@ -85,7 +84,7 @@
           >
 
           <a
-            href="api/games"
+            href="/game/all"
             class="text-gray-300 hover:bg-gray-800 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
             >Winners</a
           >
@@ -134,9 +133,6 @@
         </div>
       </div>
     </div>
-
-    <!-- </div> -->
-
     <!-- <div class="flex-shrink-0 flex items-center justify-center py-6">
       <img
         class="block lg:hidden h-40 w-40"
@@ -148,6 +144,6 @@
         src="https://res.cloudinary.com/adev48/image/upload/v1652573714/Sofka%20Assets/Semana%204/dices_tzeuva.png"
         alt="Workflow"
       />
-    </div> -->  
+    </div>  -->
   </body>
 </html>
