+++ DiceGame/views/template/headHTML.ejs
@@ -32,7 +32,7 @@
         <div class="hidden sm:block sm:ml-6">
           <div class="flex space-x-4">
             <a
-              href="api/createGame"
+              href="/createGame"
               class="text-white hover:bg-green-800 hover:text-white px-10 py-6 rounded-md text-sm font-medium"
               aria-current="page"
               >Create Game</a
@@ -51,7 +51,7 @@
             >
 
             <a
-              href="api/games"
+              href="/game/all"
               class="text-white hover:bg-green-800 hover:text-white px-10 py-6 rounded-md text-sm font-medium"
               >Games</a
             >
@@ -66,7 +66,7 @@
     <div class="px-2 pt-2 pb-3 space-y-1">
       <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" -->
       <a
-        href="api/createGame"
+        href="/createGame"
         class="text-gray-300 hover:bg-gray-800 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
         aria-current="page"
         >Create Game</a
@@ -86,7 +86,7 @@
       >
 
       <a
-        href="api/games"
+        href="/game/all"
         class="text-gray-300 hover:bg-gray-800 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
         >Games</a
       >
